Extract code generator interface and default implementation
