+++ VaadinatorTooling/VaadinatorGenerator/src/main/java/de/akquinet/engineering/vaadinator/generator/CodeGenerator.java
@@ -1,0 +1,7 @@
+package de.akquinet.engineering.vaadinator.generator;
+
+import de.akquinet.engineering.vaadinator.mojo.VaadinatorConfig;
+
+public interface CodeGenerator {
+	public void generateCode(VaadinatorConfig vaadinatorConfig) throws Exception;
+}
